+++ app/Http/Controllers/CongePriseController.php
@@ -9,6 +9,10 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use App\Http\Requests\CongePriseRequest;
+use App\Http\Requests\GetCongeRestanteRequest;
+use App\Http\Requests\GetYearRequest;
+use App\Http\Requests\ValideCongeRequest;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class CongePriseController extends Controller
 {
@@ -28,10 +32,20 @@ class CongePriseController extends Controller
                     return $row->personnel->nom . ' ' . $row->personnel->prenom;
                 })
                 ->addColumn('action', function ($row) {
-                    if (Auth::user()->hasAnyRole('Ressource Humaine', 'Super Admin')) {
-                        $btnEditer = '<button class="btn btn-warning btn-sm mb-3" onclick="openCongePriseModal(\'edit\', ' . $row->id . ')">Éditer</button>';
-                        $btnSupprimer = '<button class="btn btn-danger btn-sm mb-3" onclick="deleteCongePrise(' . $row->id . ')">Supprimer</button>';
-                        return $btnEditer . ' ' . $btnSupprimer;
+                    if ($row->status == 'stand by') {
+                        if (Auth::user()->hasAnyRole(['SSE', 'SMF', 'SPSS', 'Chefferie', 'Super Admin'])) {
+                            $btnValider = '<button class="btn btn-success btn-sm mb-3" onclick="validerConge(' . $row->id . ')"><i class="fas fa-circle-check"></i></button>';
+                        } else {
+                            $btnValider = '';
+                        }
+                        if (Auth::user()->hasAnyRole('Ressource Humaine', 'Super Admin')) {
+                            $btnEditer = '<button class="btn btn-warning btn-sm mb-3" onclick="openCongePriseModal(\'edit\', ' . $row->id . ')"><i class="fas fa-pen"></i></button>';
+                            $btnSupprimer = '<button class="btn btn-danger btn-sm mb-3" onclick="deleteCongePrise(' . $row->id . ')"><i class="fas fa-trash"></i></button>';
+                        } else {
+                            $btnEditer = '';
+                            $btnSupprimer = '';
+                        }
+                        return $btnValider . ' ' . $btnEditer . ' ' . $btnSupprimer;
                     }
                 })
                 ->rawColumns(['action'])
@@ -49,24 +63,46 @@ class CongePriseController extends Controller
         return view('conge-prises.index', compact('mainSegment', 'personnels'));
     }
 
+    public function valideConge(ValideCongeRequest $request)
+    {
+        if (Auth::user()->hasAnyRole(['SSE', 'SPSS', 'SMF', 'Chefferie', 'Super Admin'])) {
+            $validatedData = $request->validated();
+            $congePrise = CongePrise::find($validatedData['conge_prise_id']);
+            $congePrise->update([
+                'status' => 'valide',
+            ]);
+            return response()->json([
+                'success' => true,
+                'message' => 'Congé valide',
+                'congePrise' => $congePrise
+            ]);
+        } else {
+            throw new AuthorizationException('Vous n\'etes pas autorisé à accéder à cette ressource.');
+        }
+    }
+
     /**
      * Récupère les années associées à un personnel pour mise à jour de la liste déroulante.
      *
      * @param  int  $personnelId
      * @return \Illuminate\Http\JsonResponse
      */
-    public function getYears($personnelId)
+    public function getYears(GetYearRequest $request)
     {
-        // Récupère les années associées au personnel depuis la base de données
-        $years = CongeCumule::where('personnel_id', $personnelId)->pluck('annee');
+        if (Auth::user()->hasAnyRole('Ressource Humaine', 'Super Admin')) {
+            $validatedData = $request->validated();
+            // Récupère les années associées au personnel depuis la base de données
+            $years = CongeCumule::where('personnel_id', $validatedData['personnel_id'])->pluck('annee');
+            // Crée une chaîne HTML pour les options de la liste déroulante
+            $options = '<option value="">Veuillez choisir</option>';
+            foreach ($years as $year) {
+                $options .= '<option value="' . $year . '">' . $year . '</option>';
+            }
 
-        // Crée une chaîne HTML pour les options de la liste déroulante
-        $options = '<option value="">Veuillez choisir</option>';
-        foreach ($years as $year) {
-            $options .= '<option value="' . $year . '">' . $year . '</option>';
+            return $options;
+        } else {
+            throw new AuthorizationException('Vous n\'êtes pas autorisé à accéder à cette ressource.');
         }
-
-        return $options;
     }
 
     /**
@@ -76,14 +112,22 @@ class CongePriseController extends Controller
      * @param  int  $annee
      * @return \Illuminate\Http\JsonResponse
      */
-    public function getCongeRestante($personnelId, $annee)
+    public function getCongeRestante(GetCongeRestanteRequest $request)
     {
-        // Récupère le congé restant associé au personnel et à l'année depuis la base de données
-        $congeRestante = CongeCumule::where('personnel_id', $personnelId)
-            ->where('annee', $annee)
-            ->pluck('jour_reste');
+        if (Auth::user()->hasAnyRole('Ressource Humaine', 'Super Admin')) {
+            // Retrieve the validated data
+            $validatedData = $request->validated();
 
-        return response()->json($congeRestante);
+            // Retrieve the remaining leave associated with the personnel and year from the database
+            $congeRestante = CongeCumule::where('personnel_id', $validatedData['personnel_id'])
+                ->where('annee', $validatedData['annee'])
+                ->pluck('jour_reste');
+            // Récupère le congé restant associé au personnel et à l'année depuis la base de données
+
+            return response()->json($congeRestante);
+        } else {
+            throw new AuthorizationException('Vous n\'êtes pas autorisé à accéder à cette ressource.');
+        }
     }
 
     /**
@@ -91,47 +135,51 @@ class CongePriseController extends Controller
      */
     public function store(CongePriseRequest $request)
     {
-        try {
-            // Valider les données de la requête
-            $validatedData = $request->validated();
+        if (Auth::user()->hasAnyRole('Ressource Humaine', 'Super Admin')) {
+            try {
+                // Valider les données de la requête
+                $validatedData = $request->validated();
 
-            // Calculer le nombre de jours pris à partir de la table conge_cumule
-            $congeCumule = CongeCumule::where('personnel_id', $validatedData['personnel_id'])
-                ->where('annee', $validatedData['annee'])
-                ->firstOrFail();
+                // Calculer le nombre de jours pris à partir de la table conge_cumule
+                $congeCumule = CongeCumule::where('personnel_id', $validatedData['personnel_id'])
+                    ->where('annee', $validatedData['annee'])
+                    ->firstOrFail();
 
-            $jourPrise = $congeCumule->jour_prise + $validatedData['nombre_jour'];
+                $jourPrise = $congeCumule->jour_prise + $validatedData['nombre_jour'];
 
-            // Vérifier si le nombre de jours pris dépasse le nombre total de jours disponibles
-            if ($jourPrise > $congeCumule->jour_total) {
+                // Vérifier si le nombre de jours pris dépasse le nombre total de jours disponibles
+                if ($jourPrise > $congeCumule->jour_total) {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'Le nombre de jours pris dépasse le nombre total de jours disponibles.',
+                    ], 422);
+                }
+
+                // Créer une nouvelle instance de la prise de congé
+                $newPriseConge = CongePrise::create($validatedData);
+
+                // Mettre à jour le nombre de jours pris dans la table conge_cumule
+                $congeCumule->update([
+                    'jour_prise' => $jourPrise,
+                    'jour_reste' => $congeCumule->jour_total - $jourPrise,
+                ]);
+
+                // Retourner une réponse de succès avec la nouvelle prise de congé
+                return response()->json([
+                    'success' => true,
+                    'message' => 'Prise de congé enregistrée avec succès.',
+                    'data' => $newPriseConge,
+                ], 201);
+            } catch (\Exception $e) {
+                // Retourner une réponse d'erreur en cas d'exception
                 return response()->json([
                     'success' => false,
-                    'message' => 'Le nombre de jours pris dépasse le nombre total de jours disponibles.',
-                ], 422);
+                    'message' => 'Échec de l\'enregistrement de la prise de congé.',
+                    'error' => $e->getMessage(),
+                ], 500);
             }
-
-            // Créer une nouvelle instance de la prise de congé
-            $newPriseConge = CongePrise::create($validatedData);
-
-            // Mettre à jour le nombre de jours pris dans la table conge_cumule
-            $congeCumule->update([
-                'jour_prise' => $jourPrise,
-                'jour_reste' => $congeCumule->jour_total - $jourPrise,
-            ]);
-
-            // Retourner une réponse de succès avec la nouvelle prise de congé
-            return response()->json([
-                'success' => true,
-                'message' => 'Prise de congé enregistrée avec succès.',
-                'data' => $newPriseConge,
-            ], 201);
-        } catch (\Exception $e) {
-            // Retourner une réponse d'erreur en cas d'exception
-            return response()->json([
-                'success' => false,
-                'message' => 'Échec de l\'enregistrement de la prise de congé.',
-                'error' => $e->getMessage(),
-            ], 500);
+        } else {
+            throw new AuthorizationException('Vous n\'êtes pas autorisé à accéder à cette ressource.');
         }
     }
 
@@ -141,7 +189,11 @@ class CongePriseController extends Controller
      */
     public function edit(CongePrise $congePrise)
     {
-        return response()->json($congePrise);
+        if (Auth::user()->hasAnyRole('Ressource Humaine', 'Super Admin')) {
+            return response()->json($congePrise);
+        } else {
+            throw new AuthorizationException('Vous n\'êtes pas autorisé à accéder à cette ressource.');
+        }
     }
 
     /**
@@ -149,40 +201,45 @@ class CongePriseController extends Controller
      */
     public function update(CongePriseRequest $request, CongePrise $congePrise)
     {
-        try {
-            // Valider les données de la requête
-            $validatedData = $request->validate();
+        if (Auth::user()->hasAnyRole('Ressource Humaine', 'Super Admin')) {
 
-            // Mettre à jour la prise de congé
-            $congePrise->update($validatedData);
+            try {
+                // Valider les données de la requête
+                $validatedData = $request->validate();
 
-            // Mettre à jour le nombre de jours pris dans la table conge_cumule
-            $congeCumule = CongeCumule::where('personnel_id', $congePrise->personnel_id)
-                ->where('annee', $congePrise->annee)
-                ->firstOrFail();
+                // Mettre à jour la prise de congé
+                $congePrise->update($validatedData);
 
-            // Créer une nouvelle instance de la prise de congé
-            $newPriseConge = CongePrise::create($validatedData);
+                // Mettre à jour le nombre de jours pris dans la table conge_cumule
+                $congeCumule = CongeCumule::where('personnel_id', $congePrise->personnel_id)
+                    ->where('annee', $congePrise->annee)
+                    ->firstOrFail();
 
-            // Mettre à jour le nombre de jours pris dans la table conge_cumule
-            $congeCumule->update([
-                'jour_prise' => $validatedData['nombre_jour'],
-                'jour_reste' => $congeCumule->jour_total - $validatedData['nombre_jour'],
-            ]);
+                // Créer une nouvelle instance de la prise de congé
+                $newPriseConge = CongePrise::create($validatedData);
 
-            // Retourner une réponse de succès avec la prise de congé mise à jour
-            return response()->json([
-                'success' => true,
-                'message' => 'Prise de congé mise à jour avec succès.',
-                'data' => $congePrise->refresh(), // Rafraîchir les données de la prise de congé
-            ]);
-        } catch (\Exception $e) {
-            // Retourner une réponse d'erreur en cas d'exception
-            return response()->json([
-                'success' => false,
-                'message' => 'Échec de la mise à jour de la prise de congé.',
-                'error' => $e->getMessage(),
-            ], 500);
+                // Mettre à jour le nombre de jours pris dans la table conge_cumule
+                $congeCumule->update([
+                    'jour_prise' => $validatedData['nombre_jour'],
+                    'jour_reste' => $congeCumule->jour_total - $validatedData['nombre_jour'],
+                ]);
+
+                // Retourner une réponse de succès avec la prise de congé mise à jour
+                return response()->json([
+                    'success' => true,
+                    'message' => 'Prise de congé mise à jour avec succès.',
+                    'data' => $congePrise->refresh(), // Rafraîchir les données de la prise de congé
+                ]);
+            } catch (\Exception $e) {
+                // Retourner une réponse d'erreur en cas d'exception
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Échec de la mise à jour de la prise de congé.',
+                    'error' => $e->getMessage(),
+                ], 500);
+            }
+        } else {
+            throw new AuthorizationException('Vous n\'êtes pas autorisé à accéder à cette ressource.');
         }
     }
 
@@ -193,31 +250,35 @@ class CongePriseController extends Controller
      */
     public function destroy(CongePrise $congePrise)
     {
-        try {
-            // Mettre à jour le nombre de jours restants dans la table conge_cumule
-            $congeCumule = CongeCumule::where('personnel_id', $congePrise->personnel_id)
-                ->where('annee', $congePrise->annee)
-                ->firstOrFail();
+        if (Auth::user()->hasAnyRole('Ressource Humaine', 'Super Admin')) {
+            try {
+                // Mettre à jour le nombre de jours restants dans la table conge_cumule
+                $congeCumule = CongeCumule::where('personnel_id', $congePrise->personnel_id)
+                    ->where('annee', $congePrise->annee)
+                    ->firstOrFail();
 
-            $congeCumule->update([
-                'jour_reste' => $congeCumule->jour_reste + $congePrise->nombre_jour, // Ajouter les jours supprimés
-                'jour_prise' =>  $congeCumule->jour_prise - $congePrise->nombre_jour
-            ]);
-            // Supprimer la prise de congé
-            $congePrise->delete();
+                $congeCumule->update([
+                    'jour_reste' => $congeCumule->jour_reste + $congePrise->nombre_jour, // Ajouter les jours supprimés
+                    'jour_prise' =>  $congeCumule->jour_prise - $congePrise->nombre_jour
+                ]);
+                // Supprimer la prise de congé
+                $congePrise->delete();
 
-            // Retourner une réponse de succès
-            return response()->json([
-                'success' => true,
-                'message' => 'Prise de congé supprimée avec succès.',
-            ]);
-        } catch (\Exception $e) {
-            // Retourner une réponse d'erreur en cas d'exception
-            return response()->json([
-                'success' => false,
-                'message' => 'Échec de la suppression de la prise de congé.',
-                'error' => $e->getMessage(),
-            ], 500);
+                // Retourner une réponse de succès
+                return response()->json([
+                    'success' => true,
+                    'message' => 'Prise de congé supprimée avec succès.',
+                ]);
+            } catch (\Exception $e) {
+                // Retourner une réponse d'erreur en cas d'exception
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Échec de la suppression de la prise de congé.',
+                    'error' => $e->getMessage(),
+                ], 500);
+            }
+        } else {
+            throw new AuthorizationException('Vous n\'êtes pas autorisé à accéder à cette ressource.');
         }
     }
 }
